pyproject.toml
[project]
name = "Avito"
version = "0.1.0"
description = ""
authors = [
    { name = "Kazakov Ivan", email = "vanyakazakov499@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "structlog (>=25.5.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "assertpy (>=1.1,<2.0)",
    "allure-pytest (>=2.15.3,<3.0.0)",
    "pep8 (>=1.7.1,<2.0.0)",
    "pytest-rerunfailures (>=16.1,<17.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

